int2025-telegram-bot/requirements.txt
aiogram==3
requests==2.28.1
python-dotenv==0.20.0